requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
dominate==2.5.1
idna==3.2
imageio==2.9.0
jsonpatch==1.32
jsonpointer==2.1
kiwisolver==1.3.1
matplotlib==3.4.3
networkx==2.6.2
numpy==1.21.2
Pillow==6.2.2
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
PyWavelets==1.1.1
pyzmq==22.2.1
requests==2.26.0
scikit-image==0.18.2
scipy==1.7.1
six==1.16.0
tifffile==2021.8.8
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
typing-extensions==3.10.0.0
urllib3==1.26.6
imageio-ffmpeg